Added HsvThresholdTrackbar.cpp source file
