+++ src/.vscode/tasks.json
@@ -16,6 +16,7 @@
                 "./ModularClasses/GaussianBlurTrackbar.cpp",
                 "./ModularClasses/CannyThresholdTrackbar.cpp",
                 "./ModularClasses/HsvThresholdTrackbar.hpp",
+                "./ModularClasses/HsvThresholdTrackbar.cpp",
                 //"./Rectangle_Detection/RectDetect1.hpp",
                 //"./Rectangle_Detection/RectDetect1.cpp",
                 //"find_paper.hpp",
